DevUtils/requirements.txt
Package Version
------- -------
pillow  11.0.0
pip     24.3.1
ujson   5.10.0
